LayoutTests/editing/style/text-decoration-state-expected.txt

Test to make sure we return correct text-decoration state. Note that "text-decoration: none" SHOULD NOT cancel text decorations.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS textdecorationState("underline","<u><b><i><span id=e>hello world</span></i></b></u>") is true
PASS textdecorationState("underline","<b><i><u><span id=e>hello world</span></u></i></b>") is true
PASS textdecorationState("underline","<b><i><span id=e style='text-decoration: underline;'>hello world</span></i></b>") is true
PASS textdecorationState("underline","<span style='text-decoration: underline;'><em id=e>hello world</em></span>") is true
PASS textdecorationState("underline","<u><b><i><span id=e style='text-decoration:none'>hello world</span></i></b></u>") is true
PASS textdecorationState("strikeThrough","<b><i><span id=e>hello world</span></i></b>") is false
PASS textdecorationState("strikeThrough","<s><b><i><span id=e>hello world</span></i></b></s>") is true
PASS textdecorationState("strikeThrough","<b><i><s><span id=e>hello world</span></s></i></b>") is true
PASS textdecorationState("strikeThrough","<b><i><span id=e style='text-decoration: line-through;'>hello world</span></i></b>") is true
PASS textdecorationState("strikeThrough","<span style='text-decoration: line-through;'><em id=e>hello world</em></span>") is true
PASS textdecorationState("strikeThrough","<s><b><i><span id=e style='text-decoration:none'>hello world</span></i></b></s>") is true
PASS successfullyParsed is true

TEST COMPLETE
